+++ Src/MoneyFox.Presentation/ViewModels/PaymentViewModel.cs
@@ -15,7 +15,7 @@ namespace MoneyFox.Presentation.ViewModels
         private int chargedAccountId;
         private int? targetAccountId;
         private DateTime date;
-        private double amount;
+        private decimal amount;
         private bool isCleared;
         private PaymentType type;
         private string note;
@@ -81,10 +81,10 @@ namespace MoneyFox.Presentation.ViewModels
         /// <summary>
         ///     Amount of the payment. Has to be >= 0. If the amount is charged or not is based on the payment type.
         /// </summary>
-        public double Amount {
+        public decimal Amount {
             get => amount;
             set {
-                if (Math.Abs(amount - value) < 0.01) return;
+                if (Math.Abs(amount - value) < 0.01m) return;
                 amount = value;
                 RaisePropertyChanged();
             }
